watch_finder_v2/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
faiss-cpu==1.7.4
filelock==3.18.0
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.17.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.41.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.24.3
openai==1.82.1
packaging==25.0
pandas==2.2.3
Pillow==10.0.0
plotly==6.1.2
pluggy==1.6.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.30.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
